# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Smart Building API
  description: Sistema de controle inteligente de ar-condicionado
  version: 1.0.0
paths:
...
  /api/v1/devices:
    get:
      tags:
        - Devices
      summary: Listar dispositivos
      operationId: list_devices_api_v1_devices_get
      security:
        - HTTPBearer: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            title: Page
        - name: size
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 20
            title: Size
        - name: room_id
          in: query
          required: false
          schema:
            anyOf:
              - type: string
                format: uuid
              - type: 'null'
            title: Room Id
        - name: status
          in: query
          required: false
          schema:
            anyOf:
              - $ref: '#/components/schemas/DeviceStatus'
              - type: 'null'
            title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response List Devices Api V1 Devices Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
        - Devices
      summary: Cadastrar dispositivo
      operationId: create_device_api_v1_devices_post
      security:
        - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  '/api/v1/devices/{device_id}':
    get:
      tags:
        - Devices
      summary: Detalhes de um dispositivo
      operationId: get_device_api_v1_devices__device_id__get
      security:
        - HTTPBearer: []
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Device Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
        - Devices
      summary: Atualizar dispositivo
      operationId: update_device_api_v1_devices__device_id__patch
      security:
        - HTTPBearer: []
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Device Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
        - Devices
      summary: Desativar dispositivo
      operationId: delete_device_api_v1_devices__device_id__delete
      security:
        - HTTPBearer: []
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Device Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  '/api/v1/devices/{device_id}/control':
    post:
      tags:
        - Devices
      summary: Controlar dispositivo (ligar/desligar/setpoint)
      operationId: control_device_api_v1_devices__device_id__control_post
      security:
        - HTTPBearer: []
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Device Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceControlRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceControlResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  '/api/v1/devices/{device_id}/status':
    get:
      tags:
        - Devices
      summary: Status atual do dispositivo
      operationId: get_device_status_api_v1_devices__device_id__status_get
      security:
        - HTTPBearer: []
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Device Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/predictions/24h:
    get:
      tags:
        - Predictions
      summary: Predição de consumo para as próximas 24 horas
      operationId: get_predictions_24h_api_v1_predictions_24h_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Get Predictions 24H Api V1 Predictions 24H Get
      security:
        - HTTPBearer: []
  /api/v1/predictions/train:
    post:
      tags:
        - Predictions
      summary: Retreinar modelo com dados históricos
      description: Agenda retreinamento do modelo ML com últimos 30 dias de dados.
      operationId: trigger_retrain_api_v1_predictions_train_post
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Trigger Retrain Api V1 Predictions Train Post
      security:
        - HTTPBearer: []
...
components:
  schemas:
...
    DeviceControlRequest:
      properties:
        action:
          type: string
          pattern: ^(on|off|setpoint)$
          title: Action
          description: 'Ação: ''on'', ''off'' ou ''setpoint'''
        value:
          anyOf:
            - type: number
              maximum: 30
              minimum: 16
            - type: 'null'
          title: Value
          description: Temperatura desejada (obrigatório para setpoint)
      type: object
      required:
        - action
      title: DeviceControlRequest
    DeviceControlResponse:
      properties:
        status:
          type: string
          title: Status
        device_id:
          type: string
          format: uuid
          title: Device Id
        action:
          type: string
          title: Action
        value:
          anyOf:
            - type: number
            - type: 'null'
          title: Value
        timestamp:
          type: string
          format: date-time
          title: Timestamp
      type: object
      required:
        - status
        - device_id
        - action
        - timestamp
      title: DeviceControlResponse
    DeviceCreate:
      properties:
        room_id:
          type: string
          format: uuid
          title: Room Id
        device_type:
          $ref: '#/components/schemas/DeviceType'
        model:
          anyOf:
            - type: string
            - type: 'null'
          title: Model
        serial_number:
          anyOf:
            - type: string
            - type: 'null'
          title: Serial Number
        setpoint_celsius:
          anyOf:
            - type: number
              maximum: 30
              minimum: 16
            - type: 'null'
          title: Setpoint Celsius
        mqtt_topic:
          anyOf:
            - type: string
            - type: 'null'
          title: Mqtt Topic
        notes:
          anyOf:
            - type: string
            - type: 'null'
          title: Notes
      type: object
      required:
        - room_id
        - device_type
      title: DeviceCreate
    DeviceResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        room_id:
          type: string
          format: uuid
          title: Room Id
        device_type:
          $ref: '#/components/schemas/DeviceType'
        model:
          anyOf:
            - type: string
            - type: 'null'
          title: Model
        serial_number:
          anyOf:
            - type: string
            - type: 'null'
          title: Serial Number
        status:
          $ref: '#/components/schemas/DeviceStatus'
        is_active:
          type: boolean
          title: Is Active
        power_on:
          type: boolean
          title: Power On
        setpoint_celsius:
          anyOf:
            - type: number
            - type: 'null'
          title: Setpoint Celsius
        mqtt_topic:
          anyOf:
            - type: string
            - type: 'null'
          title: Mqtt Topic
        notes:
          anyOf:
            - type: string
            - type: 'null'
          title: Notes
        last_seen_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Last Seen At
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Updated At
        links:
          items:
            $ref: '#/components/schemas/HateoasLink'
          type: array
          title: Links
      type: object
      required:
        - id
        - room_id
        - device_type
        - status
        - is_active
        - power_on
        - created_at
      title: DeviceResponse
    DeviceStatus:
      type: string
      enum:
        - online
        - offline
        - maintenance
      title: DeviceStatus
    DeviceType:
      type: string
      enum:
        - ac
        - temperature_sensor
        - humidity_sensor
        - presence_sensor
        - window_sensor
      title: DeviceType
    DeviceUpdate:
      properties:
        model:
          anyOf:
            - type: string
            - type: 'null'
          title: Model
        setpoint_celsius:
          anyOf:
            - type: number
              maximum: 30
              minimum: 16
            - type: 'null'
          title: Setpoint Celsius
        mqtt_topic:
          anyOf:
            - type: string
            - type: 'null'
          title: Mqtt Topic
        notes:
          anyOf:
            - type: string
            - type: 'null'
          title: Notes
        is_active:
          anyOf:
            - type: boolean
            - type: 'null'
          title: Is Active
      type: object
      title: DeviceUpdate
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HateoasLink:
      properties:
        rel:
          type: string
          title: Rel
        href:
          type: string
          title: Href
        method:
          type: string
          title: Method
          default: GET
      type: object
      required:
        - rel
        - href
      title: HateoasLink
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
...